Limit http/set_metadata source files to one logger ID
